refactor(serialization): extract DagDependency as a separate module
